requirements.txt
Django==2.0.1
django-ckeditor==5.4.0
django-decouple==2.1
django-js-asset==0.1.1
gunicorn==19.7.1
olefile==0.44
Pillow==4.0.0
psycopg2==2.7.1
pytz==2017.3
